TG-70: still fixing routes, endpoints ready to connect
