+++ src/Components/Tasks/List/Row/row.module.css
@@ -1,6 +1,3 @@
-.container{
-    background-color: rgb(226, 225, 225);
-}
 .row {
     border-color: rgb(150, 81, 81);
     border-style: groove;
